fix: use uint8 dtype for torch.save pickle data in training demo
torch.save produces pickle-formatted bytes with headers and metadata,
not raw bf16 tensor data. Using "bf16" dtype caused a panic in byte
grouping due to odd-length data. Switch to "uint8" to skip byte
grouping while still benefiting from delta encoding.

Also adds real_checkpoint_demo video files (.cast, .gif, .mp4).

+++ examples/training_demo.py
@@ -130,8 +130,10 @@ def main():
         raw_size = len(raw_bytes)
 
         # Compress with Mithril (uses delta from previous step automatically)
+        # Use "uint8" since torch.save produces pickle-formatted bytes, not raw tensors.
+        # Delta encoding still works great — consecutive state_dicts share most bytes.
         compressed, stats = compressor.compress_checkpoint(
-            f"step_{step}", raw_bytes, "bf16"
+            f"step_{step}", raw_bytes, "uint8"
         )
         compressed_size = len(compressed)
